feat: collect logs when report issue
