+++ desktop/src/pages/settings/viewModel.ts
@@ -4,13 +4,14 @@ import * as shell from '@tauri-apps/plugin-shell'
 import { useEffect, useRef, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import * as config from '~/lib/config'
-import { NamedPath, getAppInfo, getIssueUrl, getPrettyVersion, ls, resetApp } from '~/lib/utils'
+import { NamedPath, getIssueUrl, ls, resetApp } from '~/lib/utils'
 import { usePreferenceProvider } from '~/providers/Preference'
 import { UnlistenFn, listen } from '@tauri-apps/api/event'
 import { useNavigate } from 'react-router-dom'
 import { load } from '@tauri-apps/plugin-store'
 import { useStoreValue } from '~/lib/useStoreValue'
 import * as clipboard from '@tauri-apps/plugin-clipboard-manager'
+import { collectLogs, getPrettyVersion } from '~/lib/logs'
 
 async function openModelPath() {
 	let dst = await invoke<string>('get_models_folder')
@@ -23,7 +24,32 @@ async function openModelsUrl() {
 
 async function reportIssue() {
 	try {
-		const info = await getAppInfo()
+		let info = await collectLogs()
+
+		const logs: string = await invoke<string>('get_logs')
+		const filteredLogs = logs
+			.split('\n')
+			.filter((l) => l.toLowerCase().includes('error')) // Filter lines with "debug"
+			.slice(-10) // Take the last 3 lines
+			.map((line) => {
+				try {
+					console.log('line => ', line)
+					const parsed = JSON.parse(line) // Deserialize JSON
+					return parsed?.fields?.message || 'No message found' // Extract .message or fallback
+				} catch (e) {
+					return 'Invalid JSON' // Handle invalid JSON
+				}
+			})
+			.join('\n')
+		const templatedLogs = `<details>
+<summary>logs</summary>
+
+\`\`\`console
+${filteredLogs}
+\`\`\`
+</details>
+`
+		info += `\n\n\n${templatedLogs}`
 		shell.open(await getIssueUrl(info))
 	} catch (e) {
 		console.error(e)
@@ -130,8 +156,6 @@ export function viewModel() {
 			listenersRef.current.forEach((unlisten) => unlisten())
 		}
 	}, [])
-
-	useEffect(() => { }, [preference])
 
 	return {
 		copyLogs,
